pyproject.toml
[project]
name = "packagerix"
version = "0.1.0"
description = "LLM-assisted packaging with Nix"
authors = [{name = "Martin Schwaighofer", email = "mschwaig+packagerix@eml.cc"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "setuptools",
    "wheel",
    "textual",
    "rich",
    "requests",
    "beautifulsoup4",
    "diskcache",
    "pydantic",
    "GitPython",
    "magentic",
    "litellm",
    "loguru",
    "keyring",
    "keyrings.alt",
]

[project.scripts]
packagerix = "packagerix.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
packagerix = ["template/*"]
